Move parsersJSFacade into parsers subproject
