+++ build.sbt
@@ -145,7 +145,8 @@ lazy val parsers = crossProject
   .in(file("scalameta/parsers"))
   .settings(
     publishableSettings,
-    description := "Scalameta APIs for parsing and their baseline implementation"
+    description := "Scalameta APIs for parsing and their baseline implementation",
+    scalaJSModuleKind := ModuleKind.CommonJSModule
   )
   .dependsOn(common, dialects, inputs, tokens, tokenizers, trees)
 lazy val parsersJVM = parsers.jvm
@@ -683,12 +684,3 @@ def CiCommand(name: String)(commands: List[String]): Command = Command.command(n
   }
 }
 def ci(command: String) = s"plz $ciScalaVersion $command"
-
-lazy val parsersJSFacade = project
-  .in(file("scalameta/parsersJSFacade"))
-  .enablePlugins(ScalaJSPlugin)
-  .settings(
-    sharedSettings,
-    scalaJSModuleKind := ModuleKind.CommonJSModule
-  )
-  .dependsOn(parsersJS)
